Refactor out LegacyButton behaviour
